% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData.R
\docType{data}
\name{Alu.demo}
\alias{Alu.demo}
\title{Subset of Alu sequence dataset}
\format{A \code{GRanges} object.}
\source{
RepeatMasker database provided by package \code{AnnotationHub}.
}
\usage{
Alu.demo
}
\value{
A GRanges object with 500 ranges and 3 metadata columns.
}
\description{
A \code{GRanges} dataset containing 500 Alu sequences that have CpGs profiled 
by both Illumina 450k and EPIC array. The variables are as follows:
}
\details{
\itemize{
  \item seqnames chromosome number
  \item ranges hg19 genomic position
  \item strand DNA strand
  \item name Alu subfamily
  \item score Smith Waterman (SW) alignment score
  \item Index internal index
}

Scripts for generating this object are contained in the \code{scripts} directory.
}
\keyword{datasets}
