% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData.R
\docType{data}
\name{Alu.hg19.demo}
\alias{Alu.hg19.demo}
\title{Subset of Alu genomic location dataset (hg19)}
\format{
A \code{\link{GRanges}} object.
}
\source{
RepeatMasker database provided by package \code{AnnotationHub}
}
\usage{
Alu.hg19.demo
}
\value{
A GRanges object with 500 ranges and 3 metadata columns.
}
\description{
A \code{\link{GRanges}} dataset containing 500 Alu sequences that have CpGs profiled
by both Illumina 450k and EPIC array. The variables are as follows:
}
\details{
\itemize{
  \item seqnames: chromosome number
  \item ranges: hg19 genomic position
  \item strand: DNA strand
  \item name: Alu subfamily
  \item score: Smith Waterman (SW) alignment score
  \item Index: internal index (meaningless for external use; 
  not communicable between genome builds)
}

\code{Alu.hg19.demo} has the same format as the data object 
returned by \code{\link{fetchRMSK}}.
}
\seealso{
See \code{\link{fetchRMSK}} to obtain the complete Alu/L1 dataset.
}
\keyword{datasets}
