% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\docType{data}
\name{artists}
\alias{artists}
\title{Interpreten beim Musik Streaming}
\format{
Ein tibble mit 50 Zeilen and 4 Spalten:
\describe{
  \item{Interpret}{Name des Interpreten oder der Band.}
  \item{Genre}{Diverse Genre Bezeichnungen pro Interpreten.}
  \item{Follower}{Follower des Interpreten innerhalb der Streaming Plattform.}
  \item{Bekanntheit}{Index zur Quantifizierung der Bekanntheit des Interpreten.}
}
}
\source{
\url{Research}
}
\usage{
artists
}
\description{
Dies ist ein Datensatz mit den 50 am meisten gehörten Interpreten eines Nutzers auf einer der bekanntesten Musik Streaming Plattformen.
}
\keyword{datasets}
