% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\docType{data}
\name{B}
\alias{B}
\title{B}
\format{
Ein tibble mit 1,493 Zeilen and 4 Spalten:
\describe{
  \item{Titel}{Names des Filmes oder der Serie}
  \item{Staffel}{Staffelnummer}
  \item{Folge}{Name der Folge}
  \item{Watchdate}{Datum, an dem der Titel geschaut wurde. Beachte, dass aus didaktischen Gründen bewusst das falsche Datenformat vorliegt.}
}
}
\source{
\url{Research}
}
\usage{
B
}
\description{
TBA
}
\keyword{datasets}
