% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5_long.R
\docType{data}
\name{big5_long}
\alias{big5_long}
\title{BIG 5 im langen Format}
\format{
Ein tible mit 200 Zeilen and 16 Spalten:
\describe{
  \item{Alter}{Alter in Lebensjahren}
  \item{Geschlecht}{f = Weiblich, m = Männlich}
  \item{Gruppe}{Altersgruppe unterteilt in Jung, Mittel, Weise}
  \item{ID}{Individuelle Personenidentifikation}
  \item{Faktor}{Persönlichkeitsfaktor Extraversion oder Neurotizismus}
  \item{Auspraegung}{Mittlere Ausprägung des jeweiligen Persönlichkeitsfaktors}
  \item{Zeitpunkt}{Messzeitpunkt 1 oder 2}
}
}
\source{
\url{https://openpsychometrics.org/_rawdata/} (modifiziert)
}
\usage{
big5_long
}
\description{
Der Datensatz enthält einen Teil des Big5 Datensatzes im langen Format.
}
\examples{
big5_long

}
\keyword{datasets}
