% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5_mod2.R
\docType{data}
\name{big5_mod2}
\alias{big5_mod2}
\title{BIG 5 Persönlichkeitstest: Modifikation 2}
\format{
Ein tible mit 200 Zeilen and 16 Spalten:
\describe{
  \item{Alter}{Alter in Lebensjahren}
  \item{Geschlecht}{f = Weiblich, m = Männlich}
  \item{Extraversion}{Durchschnittliche Ausprägung von Extraversion}
  \item{Neurotizismus}{Durchschnittliche Ausprägung von Neurotizismus}
  \item{Vertraeglichkeit}{Durchschnittliche Ausprägung von Verträglichkeit}
  \item{Gewissenhaftigkeit}{Durchschnittliche Ausprägung von Gewissenhaftigkeit}
  \item{O1}{Ich besitze einen großen Wortschatz.}
  \item{O2}{Ich habe Schwierigkeiten abstrakte Ideen zu verstehen. (INVERS)}
  \item{O3}{Ich habe eine lebhafte Vorstellungskraft.}
  \item{O4}{Ich bin an abstrakten Ideen nicht interessiert. (INVERS)}
  \item{O5}{Ich habe ausgezeichnete Ideen.}
  \item{O6}{Ich habe kein gutes Vorstellungsvermögen. (INVERS)}
  \item{O7}{Ich verstehe Dinge schnell.}
  \item{O8}{Ich gebrauche schwierige Wörter.}
  \item{O9}{Ich verbringe Zeit damit über Dinge nachzudenken.}
  \item{O10}{Ich bin voller Ideen.	}
}
}
\source{
\url{https://openpsychometrics.org/_rawdata/} (modifiziert)
}
\usage{
big5_mod2
}
\description{
Der Datensatz enthält die soziodemographische Variablen Alter und Geschlecht, die durchschnittliche Ausprägung von Extraversion,
Neurotizismus, Verträglichkeit und Gewissenhaftigkeit sowie 10 Items, um die Offenheit zu messen. Für mehr Informationen siehe den Datensatz `big5_comp`.
}
\examples{
big5_mod2

}
\keyword{datasets}
