% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bildung.R
\docType{data}
\name{bildung}
\alias{bildung}
\title{Statistische Bildung}
\format{
Ein tibble mit 146 Zeilen and 71 Spalten:
\describe{
  \item{ID}{Versuchspersonennummer}
  \item{Version}{Fragebogenversion mit Intervetion und ohne}
  \item{Geschlecht}{Biologisches Geschlecht mit f:weiblich, m:männlich.}
  \item{Alter}{Alter in Lebensjahren}
  \item{Schulbildung}{Höchste abgeschlossene Schulausbildung.}
  \item{Risikosportarten}{Praktizieren mindestens einer Risikosportart.}
  \item{Statistische_Kenntnisse}{Selbsteinschätzung der eigenen Statistikkenntnisse auf einer Skala von 0 *keine Kenntnisse* bis 6 *ausgezeichnete Kenntnisse*.}
  \item{Interesse_Wissenschaft}{Subjektives Interesse an Wissenschaft auf einer Skala von 0 *kein Interesse* bis 6 *sehr starkes Interesse*.}
  \item{Dospert_verhalten}{Median der DOSPERT Skala zum Risikoverhalten mit einer Skala von 1 *sehr unwahrscheinlich* bis 7 *sehr wahrscheinlich*.}
  \item{Dospert_wahrnehmung}{Median der DOSPERT Skala zur Risikowahrnehmung mit einer Skala von 1 *überhaupt kein Risiko* bis 7 *sehr hohes Risiko*.}
  \item{Dospert_nutzen}{Median der DOSPERT Skala zum Risikonutzen mit einer Skala von 1 *gar kein nutzen* bis 7 *sehr hoher Nutzen*.}
  \item{Bildung_pre}{Mittlere Lösungswahrscheinlichkeit über alle Fragen vor der Intervention.}
  \item{Bildung_post}{Mittlere Lösungswahrscheinlichkeit über alle Fragen nach Intervention.}
  \item{Bildung_gesamt}{Mittlere Lösungswahrscheinlichkeit insgesamt.}
}
}
\usage{
bildung
}
\description{
In dieser Studie mit 146 ProbandInnen wurde zum einen versucht, einen Zusammenhang zwischen statistischer Bildung und Risikoverhalten zu finden und zum anderen eine Verbesserung statistischer Bildung durch eine graphische Hilfestellung zu entdecken.

Dafür wurden zwei verschiedene Fragebogenversionen ausgeteilt. In der einen gab es nach rund der Hälfte der Aufgaben eine graphische Hilfestellung, die eine mögliche Lösungsstrategie zu einem Teil der Aufgaben erklärt (Interventionsgruppe). Die andere Version beinhaltete lediglich eine schriftliche Erklärung.
Statistische Bildung (engl. Statistical literacy) ist als Interpretieren, kritisches Bewerten und die Fähigkeit über statistische Informationen zu sprechen definiert.
}
\examples{
bildung
}
\keyword{datasets}
