% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoin.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Bitcoin Aktienkurs}
\format{
Ein tibble mit 731 Zeilen and 2 Spalten:
\describe{
  \item{Datum}{Datum der Preismessung.}
  \item{Preis}{Preise der Aktie zu einem bestimmten Datum in Euro.}
}
}
\usage{
bitcoin
}
\description{
Preise der Bitcoin Aktie zwischen Januar 2019 und Dezember 2020.
}
\examples{
bitcoin
}
\keyword{datasets}
