% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\docType{data}
\name{covid}
\alias{covid}
\title{SARS-CoV-2 Diagnostik}
\format{
Eine Matrix mit 2 Zeilen and 2 Spalten.
\describe{
  \item{PCR}{Positives oder negatives Ergebnis des PCR-Tests. Ein positiver Test entspricht einer sicheren Diagnose einer SARS-CoV-2 Infektion.}
  \item{Schnelltest}{Positives oder negatives Ergebnis eines Antigen-Schnelltests.}
}
}
\usage{
covid
}
\description{
Kontingenztafel des Vergleichs der diagnostischen Qualität zwischen PCR-Test und Antigen-Schnelltest.
}
\examples{
covid
}
\keyword{datasets}
