% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eigen.R
\name{data_eigen}
\alias{data_eigen}
\title{Berechnung der Eigenvalues}
\usage{
data_eigen(df)
}
\arguments{
\item{df}{Data.frame oder tibble mit numerischen Spalten.}
}
\value{
Eigenvalues und Dimensionen.
}
\description{
Die ausgegebenen Werte können direkt als Scree Plot visualisiert werden.
}
\examples{
big_five_comp

big_five_comp |> 
 dplyr::select(-Geschlecht) |> 
 data_eigen()
}
