% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr.R
\docType{data}
\name{demogr}
\alias{demogr}
\title{Infekte bei Blasenstörung: Demographische Daten}
\format{
Ein tibble mit 4 Zeilen and 3 Spalten:
\describe{
  \item{ID}{TeilnehmerInnen Identifikator.}
  \item{Sex}{Biologisches Geschlecht (Auswahl) mit f:Frauen und m:Männer}
  \item{Alter}{Alter in Lebensjahrenahren.}
}
}
\usage{
demogr
}
\description{
Ein minimales Beispiel für einen Datensatz mit den demographischen Daten in Form von Geschlecht und Alter.
}
\examples{
demogr
}
\keyword{datasets}
