% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_oben.R
\docType{data}
\name{df_oben}
\alias{df_oben}
\title{df_oben}
\format{
Ein tibble mit 3 Zeilen and 4 Spalten:
\describe{
  \item{ID}{TeilnehmerInnen Identifikator.}
  \item{Puls}{Anzahl an Schlägen pro Minute.}
  \item{Blasenstoerung}{Es liegt eine Blasenstörung vor (1) oder nicht (0).}
  \item{Infekt_2j}{Anzahl an Infektionen innerhalb von 2 Jahren.}
}
}
\usage{
df_oben
}
\description{
TBA
}
\examples{
df_oben
}
\keyword{datasets}
