% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\docType{data}
\name{eye_tracking}
\alias{eye_tracking}
\title{Eye Tracking}
\format{
Ein tibble mit 100 Zeilen and 3 Spalten:
\describe{
  \item{Gesichter}{Anzahl fixierter Gesichter.}
  \item{Dichte}{Gezeigte Informationsdichte.}
  \item{Studie}{Study 1 oder Study 2.}
}
}
\usage{
eye_tracking
}
\description{
Vergleich von zwei Eye Tracking Studien zur Untersuchung der Anzahl fixierter Gesichter in Abhängigkeit der Informationsdichte
}
\examples{
eye_tracking
}
\keyword{datasets}
