% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fetchRMSK}
\alias{fetchRMSK}
\title{Get RE database from RepeatMasker}
\usage{
fetchRMSK(ah, REtype, verbose = FALSE)
}
\arguments{
\item{ah}{An \code{\link{AnnotationHub}} object. Use \code{AnnotationHub()} to 
retrive information about all records in the hub.}

\item{REtype}{Type of RE. Currently \code{"Alu"} and \code{"L1"} are supported.}

\item{verbose}{Logical parameter. Should the function be verbose?}
}
\value{
A \code{\link{GRanges}} object containing RE database. 'name' column 
indicates the RE subfamily; 'score' 
column indicate the SW score; 'Index' is a interal index for RE to faciliate data 
referral, which is meaningless for external use.
}
\description{
\code{fetchRMSK} is used to obtain specified RE database from RepeatMasker Databse 
provided by AnnotationHub.
}
\examples{
ah <- AnnotationHub::AnnotationHub()
L1 <- fetchRMSK(ah, 'L1', verbose = TRUE)
L1

}

