% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hands_on.R
\name{hands_on}
\alias{hands_on}
\title{Übungen starten}
\usage{
hands_on(name)
}
\arguments{
\item{name}{Der Name der Übung als Character.}
}
\value{
Die ausgewählte Übung wird im Browser gestartet.
}
\description{
Eine komfortable Funktion zum Starten der im `remp` enthaltenen Übungen im eigenen Standardbrowser.
Es ist keine Internetverbindung notwendig.
}
\examples{
\dontrun{
hands_on("ue_datatypes")
}
}
