% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indonesisch.R
\docType{data}
\name{indonesisch}
\alias{indonesisch}
\title{Sinnerfüllung in Indonesien}
\format{
Ein tibble mit 200 Zeilen and 8 Spalten:
\describe{
  \item{Geschlecht}{f = Weiblich, m = Männlich}
  \item{Familienstand}{ledig, verheiratet}
  \item{Alter}{Alter in Lebensjahren}
  \item{Selbsttranszendenz}{Durchschnittliche Ausprägung von Selbsttranszendenz.}
  \item{Selbstverwirklichung}{Durchschnittliche Ausprägung von Selbstverwirklichung.}
  \item{Ordnung}{Durchschnittliche Ausprägung von Ordnung.}
  \item{Wir_Wohlgefuehl}{Durchschnittliche Ausprägung von Wir-Wohlgefühl.}
  \item{Sinnerfuellung}{Durchschnittliche Ausprägung von Sinnerfüllung.}

}
}
\source{
\url{Research}
}
\usage{
indonesisch
}
\description{
Ergebnisse der Hauptdimensionen des Fragebogens zu Lebensbedeutung und Lebenssinn (LEBE) an einer indonesischen Stichprobe.
}
\keyword{datasets}
