% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq.R
\docType{data}
\name{iq}
\alias{iq}
\title{IQ und Kreativität}
\format{
Ein tibble mit 90 Zeilen and 5 Spalten:
\describe{
  \item{ID}{Personenidentifikationsnummer}
  \item{Zeitpunkt}{Messzeitpunkt}
  \item{Gruppe}{Experimentalgruppen}
  \item{IQ}{Durchschnittlicher Intelligenzquotient}
  \item{Kreativitaet}{Erfundener Score zur Messung der Kreativität}
}
}
\usage{
iq
}
\description{
Studie zur Erfassung des Zusammenhangs zwischen Intelligenz und Kreativität an verschiedenen Messzeitpunkten.
}
\examples{
iq
}
\keyword{datasets}
