% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literacy.R
\docType{data}
\name{literacy}
\alias{literacy}
\title{Statistische Bildung}
\format{
Ein tibble mit 146 Zeilen and 71 Spalten:
\describe{
  \item{ID}{Versuchspersonennummer.}
  \item{Version}{Fragebogenversion (1 und 2).}
  \item{Geschlecht}{Biologisches Geschlecht mit 1:weiblich, 2:männlich (Auswahl).}
  \item{Alter}{Lebensjahre in Jahren}
}
}
\usage{
literacy
}
\description{
TBA
}
\examples{
literacy
}
\keyword{datasets}
