% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literacy_mod.R
\docType{data}
\name{literacy_mod}
\alias{literacy_mod}
\title{Statistische Bildung (Auswahl)}
\format{
Ein tibble mit 146 Zeilen and 3 Spalten:
\describe{
  \item{Risikoverhalten}{Risikoverhalten}
  \item{Schulbildung}{Schulbildung}
  \item{Literacy_total}{Ingesamte Statistikbildung.}
}
}
\usage{
literacy_mod
}
\description{
TBA
}
\examples{
literacy_mod
}
\keyword{datasets}
