% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMP-package.R
\docType{package}
\name{REMP-package}
\alias{REMP}
\alias{REMP-package}
\title{Repetitive Element Methylation Prediction}
\description{
Machine learing-based tools to predict DNA methylation of locus-specific repetitive elements (RE) 
by learning surrounding genetic and epigenetic information. These tools provide genomewide and 
single-base resolution of DNA methylation prediction on RE that are difficult to measure using 
array-based or sequencing-based platforms, which enables epigenome-wide association study (EWAS) 
and differentially methylated region (DMR) analysis on RE.

\tabular{ll}{ Name: \tab REMP\cr Type: \tab Package\cr
Version: \tab 0.99.18\cr License: \tab GPL-3\cr }
}
\section{Overview - standard procedure}{
 
\describe{
    \item{Step 1}{Start out generating required dataset for prediction using \code{\link{initREMP}}. 
    It is recommended to save these generated data to working directory so they can be used in the future.}
    \item{Step 2}{Clean Illumina methylation dataset using \code{\link{grooMethy}}. This function 
    can help identify and fix abnormal values and automatically impute missing values, which are 
    essential for downstream prediction.}
    \item{Step 3}{Run \code{\link{remp}} to predict genome-wide locus specific RE methylation.}
    \item{Step 4}{Use the built-in accessors and utilities in \code{\link{REMProduct}} object to get or 
    refine the prediction results.}
    }
}
\author{
Yinan Zheng \email{y-zheng@northwestern.edu},
Lei Liu \email{lei.liu@northwestern.edu},       
Wei Zhang \email{wei.zhang1@northwestern.edu},
Warren Kibbe \email{warren.kibbe@nih.gov},       
Lifang Hou \email{l-hou@northwestern.edu}

Maintainer: Yinan Zheng \email{y-zheng@northwestern.edu}
}
\references{
Manuscript in review (Nucleic Acids Research)
}
\keyword{package}

