% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated.R
\docType{data}
\name{repeated}
\alias{repeated}
\title{Messwiederholungsbeispiel}
\format{
Ein tibble mit 90 Zeilen and 5 Spalten:
\describe{
  \item{iq}{Intelligenzquotient.}
  \item{kreativitaet}{Kreativität.}
  \item{zeitpunkt}{Messzeitpunkt 1 oder 2.}
  \item{gruppe}{Experimentalgruppe A, B, C oder D.}
  \item{person_id}{Versuchspersonennummer.}
}
}
\usage{
repeated
}
\description{
TBA
}
\examples{
repeated
}
\keyword{datasets}
