% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipp_wm.R
\docType{data}
\name{tipp_wm}
\alias{tipp_wm}
\title{Tipp WM: langes Format}
\format{
Ein tibble mit 384 Zeilen and 7 Spalten:
\describe{
  \item{Spieltag}{Spieltag}
  \item{Person}{TipperIn}
  \item{Tipp}{Tipp}
  \item{Ergebnis}{Wahres Ergebnis}
  \item{Tipp_Richtung}{Getippte Richtung: Sieg/Unenentschiedn/Niederlage(S/N/U)}
  \item{Wahre_Richtung}{Wahre Richtung: Sieg/Unenentschiedn/Niederlage(S/N/U)}
  \item{Art}{Art des Spieles}
}
}
\usage{
tipp_wm
}
\description{
Tipps von sechs verschiedenen Personen zur Fußball Weltmeisterschaft 2018 in langem Format.
}
\examples{
tipp_wm
}
\keyword{datasets}
