% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{Lieder eines Musik Streamingdienstes}
\format{
Ein tibble mit 50 Zeilen and 6 Spalten:
\describe{
  \item{Lied}{Name des Liedes}
  \item{Interpret}{Name des Interpreten}
  \item{Album}{Name des Albums}
  \item{Anzahl}{Anzahl der Lieder auf dem jeweiligen Album}
  \item{Bekanntheit}{Index zur Quantifizierung der Bekanntheit.}
  \item{Release}{Release Datum des jeweiligen Albums/Liedes}
}
}
\usage{
tracks
}
\description{
Dies ist ein Datensatz mit den 50 am meisten gehörten Liedern eines Nutzers auf einer der bekanntesten Musik Streaming Plattformen.
}
\examples{
tracks
}
\keyword{datasets}
