% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_binary.R
\name{transform_binary}
\alias{transform_binary}
\title{Binary Transformation}
\usage{
transform_binary(df, answers)
}
\arguments{
\item{df}{Datensatz als data.frame / tibble.}

\item{answers}{Antwortvektor als Vektor / einzeiliger tibble.}
}
\value{
Ergebnismatrix als tibble
}
\description{
Erstellt eine binaere Ergebnismatrix, in der die richtigen Antworten mit einer 1 und die falschen mit einer 0 markiert sind.
Falls die richtigen Antworten sowohl Numerics als auch Characters beinhalten, kann der Funktion ebenfalls ein einzeiliger data.frame / tibble gegeben werden.
Fehlende Werte werden als falsch beantwortet gewertet.
}
\examples{
\dontrun{
########################
# Identische Datentypen
soscisurvey <- data.frame(
 I1 = c(3, 2, 3, 1),
 I2 = c(2, 2, 2, 1),
 I3 = c(4, 1, 4, 4)
)
antwortvec <- c(3, 2, 4)
transform_binary(soscisurvey, antwortvec)
########################
# Verschiedene Datentypen
soscisurvey <- data.frame(
 I1 = c(3, 2, 3, 1),
 I2 = c("Apfel", "Banane", "Erdbeere", "Apfel"),
 I3 = c(4, 1, NA, 4)
)
antwortvec <- data.frame(3, "Apfel", 4)
transform_binary(soscisurvey, antwortvec)
}
}
