% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uebung_anzeigen.R
\name{uebung_anzeigen}
\alias{uebung_anzeigen}
\title{Alle Übungen anzeigen}
\usage{
uebung_anzeigen(name = "alle")
}
\arguments{
\item{name}{Die Nummer der Übung als Character oder Numeric.}
}
\value{
Ein data.frame mit Übungsnummer und Kurzbeschreibung
}
\description{
Anzeigen der im Package enthaltenen Übungen mit Beschreibung.
}
\examples{
uebung_anzeigen()
uebung_anzeigen(6.2)
}
