% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uebung_starten.R
\name{uebung_starten}
\alias{uebung_starten}
\title{Interaktive Übung starten}
\usage{
uebung_starten(name)
}
\arguments{
\item{name}{Die Nummer der Übung als Character oder Numeric.}
}
\value{
Die ausgewählte Übung wird im Browser gestartet.
}
\description{
Starten der im Package enthaltenen Übungen innerhalb von RStudio oder des eigenen Browsers.
Dabei ist keine Internetverbindung notwendig. Der Fortschritt innerhalb der Übungen wird automatisch lokal auf dem eigenen Computer gespeichert.
}
\examples{
\dontrun{
uebung_starten("6.1")
}
}
