% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videostream.R
\docType{data}
\name{videostream}
\alias{videostream}
\title{Videostreaming Nutzerstatistik}
\format{
Ein tibble mit 1,493 Zeilen and 4 Spalten:
\describe{
  \item{Titel}{Names des Filmes oder der Serie}
  \item{Staffel}{Staffelnummer}
  \item{Folge}{Name der Folge}
  \item{Watchdate}{Datum, an dem der Titel geschaut wurde. Beachte, dass aus didaktischen Gründen bewusst das falsche Datenformat vorliegt.}
}
}
\usage{
videostream
}
\description{
Dies ist die Nutzerstatistik eines User beim Film- und Serienschauen auf einer der bekanntesten Videostreaming Plattformen.
}
\examples{
videostream
}
\keyword{datasets}
