% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5.R
\docType{data}
\name{big5}
\alias{big5}
\title{BIG 5 Modifikation}
\format{
Ein tible mit 200 Zeilen and 16 Spalten:
\describe{
  \item{Alter}{Alter in Lebensjahren}
  \item{Geschlecht}{f = Weiblich, m = Männlich}
  \item{Extraversion}{Durchschnittliche Ausprägung von Extraversion}
  \item{Neurotizismus}{Durchschnittliche Ausprägung von Neurotizismus}
  \item{O1-O3}{Item 1-3 zur Ausprägung von Offenheit}
}
}
\source{
\url{https://openpsychometrics.org/_rawdata/} (modifiziert)
}
\usage{
big5
}
\description{
Der Datensatz enthält die soziodemographische Variablen Alter und Geschlecht, die durchschnittliche Ausprägung von Extraversion,
Neurotizismus, Verträglichkeit und Gewissenhaftigkeit sowie 10 Items, um die Offenheit zu messen.
}
\examples{
big5

}
\keyword{datasets}
