% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5_zeit.R
\docType{data}
\name{big5_zeit}
\alias{big5_zeit}
\title{BIG 5 Persönlichkeitstest: Zeitpunkte und Fragebögen}
\format{
Ein tibble mit 5 Zeilen and 5 Spalten:
\describe{
  \item{VPN}{Versuchspersonennummer}
  \item{Extrav_T1}{Mittlere Extraversionsausprägung zu Zeitpunkt 1.}
  \item{Extrav_T2}{Mittlere Extraversionsausprägung zu Zeitpunkt 2.}
  \item{Neurot_FA}{Mittlere Neurotizismusausprägung in Fragebogen A.}
  \item{Neurot_FB}{Mittlere Neurotizismusausprägung in Fragebogen B.}
}
}
\usage{
big5_zeit
}
\description{
Ausschnitt aus dem BIG 5 Datensatz zur Illustrierung der Trennung von Spalten. Für mehr Informationen siehe den Datensatz big5_comp.
}
\examples{
big5_zeit
}
\keyword{datasets}
