% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5_zeit.R
\docType{data}
\name{big5_zeit}
\alias{big5_zeit}
\title{Big 5 Modifikation}
\format{
Ein tibble mit 5 Zeilen and 5 Spalten:
\describe{
  \item{VPN}{Versuchspersonennummer}
  \item{Extrav_T1}{Mittlere Extraversionsauspr채gung zu Zeitpunkt 1.}
  \item{Extrav_T2}{Mittlere Extraversionsauspr채gung zu Zeitpunkt 2.}
  \item{Neurot_FA}{Mittlere Neurotizismusauspr채gung in Fragebogen A.}
  \item{Neurot_FB}{Mittlere Neurotizismusauspr채gung in Fragebogen B.}
}
}
\usage{
big5_zeit
}
\description{
TBA
}
\examples{
big5_zeit
}
\keyword{datasets}
