% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_five_comp.R
\docType{data}
\name{big_five_comp}
\alias{big_five_comp}
\title{Big 5 Rohdatensatz}
\format{
Ein tibble mit 5000 Zeilen and 52 Spalten:
\describe{
  \item{Geschlecht}{}
  \item{E1}{Extraversion Frage 1}
  \item{N1}{Neurotizismus Frage 1}
  \item{A1}{Verträglichkeit Frage 1}
  \item{C1}{Gewissenhaftigkeit Frage 1}
  \item{O1}{Offenheit Frage 1}
}
}
\source{
\url{https://openpsychometrics.org/_rawdata/}
}
\usage{
big_five_comp
}
\description{
TBA
}
\examples{
big_five_comp

}
\keyword{datasets}
