% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoin.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Bitcoin Aktienkurs}
\format{
Ein tibble mit 731 Zeilen and 2 Spalten:
\describe{
  \item{Date}{Datum.}
  \item{Price}{Preis der Aktie in Euro.}
}
}
\usage{
bitcoin
}
\description{
TBA
}
\examples{
bitcoin
}
\keyword{datasets}
