% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eigen.R
\name{data_eigen}
\alias{data_eigen}
\title{Berechnung der Eigenwerte}
\usage{
data_eigen(df)
}
\arguments{
\item{df}{Data.frame oder tibble mit numerischen Spalten.}
}
\value{
Eigenwerte und Dimensionen.
}
\description{
Die ausgegebenen Eigenwerte können direkt als Scree Plot visualisiert werden.
}
\examples{
big5_comp

big5_comp |> 
 dplyr::select(-Geschlecht) |> 
 data_eigen()
}
