% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eigen.R
\name{data_eigen}
\alias{data_eigen}
\title{Datengenerierungsfunktion Eigenwerte}
\usage{
data_eigen(fa)
}
\arguments{
\item{fa}{Data.frame oder tibble mit numerischen Spalten.}
}
\value{
Scree Plot
}
\description{
Die Ausgegeben Werte können direkt als Scree Plot visualisiert werden.
}
\examples{
\dontrun{
data_eigen(dataset)
}
}
