% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogr.R
\docType{data}
\name{demogr}
\alias{demogr}
\title{demogr}
\format{
Ein tibble mit 4 Zeilen and 3 Spalten:
\describe{
  \item{ID}{TeilnehmerInnen Identifikator.}
  \item{Sex}{Biologisches Geschlecht (Auswahl).}
  \item{Alter}{Lebensalter in Jahren.}
}
}
\usage{
demogr
}
\description{
TBA
}
\examples{
demogr
}
\keyword{datasets}
