% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive}
\alias{descriptive}
\title{Lagemaße und Streuungsmaße berechnen}
\usage{
descriptive(data, .round = 2, .print = NA, .tibble = TRUE)
}
\arguments{
\item{data}{Daten als data.frame, tibble oder grouped.df.}

\item{.round}{Anzahl der Nachkommastellen.}

\item{.print}{Anzahl der Zeilen, die ausgegeben werden sollen. Bei Angabe einer Zahl wird das Standardverhalten eines tibbles, i.d.R nur 10 Zeilen anzuzeigen, überschrieben.}

\item{.tibble}{Ausgabe als data.frame (FALSE) oder tibble (TRUE).}
}
\value{
Überblick über typische Lage- und Streuungsmaße.
}
\description{
Berechnet die Anzahl N, Quartile, Mittelwert, Median, Standardabweichung und Standardfehler.
}
\examples{
chemo

chemo |> 
  descriptive()

chemo |> 
  dplyr::group_by(Behandlung) |> 
  descriptive(.print = 25)
}
