% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive}
\alias{descriptive}
\title{Lagemaße und Streuungsmaße berechnen}
\usage{
descriptive(data, .round = 2)
}
\arguments{
\item{data}{Daten als data.frame, tibble oder grouped.df.}

\item{.round}{Runden der Nachkommastellen. Default ist 2.}
}
\value{
Überblick über typische Lage- und Streuungsmaße.
}
\description{
Berechnet die Anzahl N, Quartile, Mittelwert, Median, Standardabweichung und Standardfehler.
}
\examples{
iq

iq |> 
  descriptive()

iq |> 
  dplyr::group_by(Zeitpunkt) |> 
  descriptive()

}
