% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive}
\alias{descriptive}
\title{Descriptive Statistics}
\usage{
descriptive(data, .round = 2)
}
\arguments{
\item{data}{Data as data.frame, tibble or grouped.df.}

\item{.round}{The digits to round to. Default is 2.}
}
\value{
Overview of common descriptive measures.
}
\description{
A function to get a quick overview over N, quartiles, mean, median, standard deviation and standard error.
}
\examples{
\dontrun{
data(iq)

iq \%>\%
  descriptive()

iq \%>\%
  group_by(Zeitpunkt) \%>\%
  descriptive()
}
}
