% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_oben.R
\docType{data}
\name{df_oben}
\alias{df_oben}
\title{Infekte bei Blasenstörung: oberer Teil}
\format{
Ein tibble mit 3 Zeilen and 4 Spalten:
\describe{
  \item{ID}{TeilnehmerInnen Identifikator.}
  \item{Puls}{Anzahl an Herzschlägen pro Minute.}
  \item{Blasenstoerung}{Es liegt eine Blasenstörung vor (1) oder nicht (0).}
  \item{Infekt_2j}{Anzahl an Infektionen innerhalb von 2 Jahren.}
}
}
\usage{
df_oben
}
\description{
Ein minimales Beispiel eines Datensatzes über die Infektionshäufigkeit bei Blasenstörung von Untersucher 1. Ein weiterer Untersucher hat ebenfalls Daten an derselben Klinik gesammelt, welche in `df_unten` gespeichert sind. Die zugehörigen demographischen Daten sind in `demogr` enthalten.
}
\examples{
df_oben
}
\keyword{datasets}
