% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_unten.R
\docType{data}
\name{df_unten}
\alias{df_unten}
\title{Infekte bei Blasenstörung: unterer Teil}
\format{
Ein tibble mit 3 Zeilen and 4 Spalten:
\describe{
  \item{ID}{TeilnehmerInnen Identifikator.}
  \item{Puls}{Anzahl an Herzschlägen pro Minute.}
  \item{Blasenstoerung}{Es liegt eine Blasenstörung vor (1) oder nicht (0).}
  \item{Infekt_2j}{Anzahl an Infektionen innerhalb von 2 Jahren.}
}
}
\usage{
df_unten
}
\description{
Ein minimales Beispiel eines Datensatzes über die Infektionshäufigkeit bei Blasenstörung von Untersucher 2. Die Daten von Untersucher 1 befinden sich in `df_oben` und die demographischen Daten in `demogr`.
}
\examples{
df_unten
}
\keyword{datasets}
