% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\docType{data}
\name{eye_tracking}
\alias{eye_tracking}
\title{Eye Tracking}
\format{
Ein tibble mit 125 Zeilen and 3 Spalten:
\describe{
  \item{Face_sum}{X.}
  \item{Density}{X.}
  \item{Gruppe}{X.}
}
}
\usage{
eye_tracking
}
\description{
TBA
}
\examples{
eye_tracking
}
\keyword{datasets}
