% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMPtools.R
\name{findRECpG}
\alias{findRECpG}
\title{Find RE-CpG genomic location given RE ranges information}
\usage{
findRECpG(RE.hg19, REtype = c("Alu", "L1"), be = NULL, verbose = FALSE)
}
\arguments{
\item{RE.hg19}{an \code{GRanges} object of RE genomic location database. This 
can be obtained by \code{\link{fetchRMSK}}.}

\item{REtype}{Type of RE. Currently \code{"Alu"} and \code{"L1"} are supported.}

\item{be}{A \code{\link{BiocParallel}} object containing back-end information that is 
ready for paralle computing. This can be obtained by \code{\link{getBackend}}.}

\item{verbose}{logical parameter. Should the function be verbose?}
}
\value{
A \code{\link{GRanges}} object containing identified RE-CpG genomic 
location data.
}
\description{
\code{findRECpG} is used to obtain RE-CpG genomic location data.
}
\details{
CpG site is defined as 5'-C-p-G-3'. It is reasonable to assume that the methylation 
status across all CpG/CpG dyads are concordant. Maintenance methyltransferase exhibits 
a preference for hemimethylated CpG/CpG dyads (methylated on one strand only). 
As a result, methyaltion status of CpG sites in both forward and reverse strands are usually consistent. 
Therefore, to accommodate the cytosine loci in both strands, the returned genomic 
ranges cover the 'CG' sequence with width of 2. The 'stand' information indicates the strand of the RE.
}
\examples{
data(Alu.demo)
be <- getBackend(1, verbose = TRUE)
RE.CpG <- findRECpG(Alu.demo, 'Alu', be, verbose = TRUE) 
RE.CpG

}
