% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hands_on.R
\name{hands_on}
\alias{hands_on}
\title{Interaktive Übung starten}
\usage{
hands_on(name)
}
\arguments{
\item{name}{Der Name der Übung als Character.}
}
\value{
Die ausgewählte Übung wird im Browser gestartet.
}
\description{
Starten der im Package enthaltenen Übungen im eigenen Browser.
Dabei ist keine Internetverbindung notwendig. Der Fortschritt innerhalb der Übungen wird automatisch lokal auf dem eigenen Computer gespeichert.
}
\examples{
\dontrun{
hands_on("ue_datatypes")
}
}
