% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literacy_mod.R
\docType{data}
\name{literacy_mod}
\alias{literacy_mod}
\title{Statistische Bildung (Auswahl)}
\format{
Ein tibble mit 1,493 Zeilen and 4 Spalten:
\describe{
  \item{Titel}{Names des Filmes oder der Serie}
  \item{Staffel}{Staffelnummer}
  \item{Folge}{Name der Folge}
  \item{Watchdate}{Datum, an dem der Titel geschaut wurde. Beachte, dass aus didaktischen Gründen bewusst das falsche Datenformat vorliegt.}
}
}
\source{
\url{Research}
}
\usage{
literacy_mod
}
\description{
TBA
}
\keyword{datasets}
