% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratios.R
\docType{data}
\name{odds_ratios}
\alias{odds_ratios}
\title{Odds Ratio Beispiel}
\format{
Ein tibble mit 3 Zeilen and 4 Spalten:
\describe{
  \item{Faktor}{Big 5 Persönlichkeitsfaktor.}
  \item{OR}{Odds Ratio.}
  \item{conf.low}{95iges Konfidenzintervall, untere Grenze.}
  \item{conf.high}{95iges Konfidenzintervall, obere Grenze.}
}
}
\usage{
odds_ratios
}
\description{
Odds Ratios ausgewählter Persönlichkeitsfaktoren aus dem Datensatz big5.
}
\examples{
odds_ratios
}
\keyword{datasets}
