% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{remp_options}
\alias{remp_options}
\alias{remp_reset}
\title{Set or get options for REMP package}
\usage{
remp_options(...)

remp_reset()
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\value{
\code{NULL}
}
\description{
Tools to manage global setting options for REMP package.
}
\section{Supported options}{

The following options are supported
\describe{
 \item{\code{.default.AluFamily}}{A list of Alu subfamily to be included in the prediction.}
 \item{\code{.default.L1Family}}{A list of L1 subfamily to be included in the prediction.}
 \item{\code{.default.GM12878.450k.URL}}{ URL to download GM12878 450k methylation profiling data.}
 \item{\code{.default.AH.repeatmasker.hg19}}{\code{AnnotationHub} data ID linked to RepeatMasker 
 database (build hg19)}
 \item{\code{.default.AH.refgene.hg19}}{\code{AnnotationHub} data ID linked to refSeq gene database 
 (build hg19)}
 \item{\code{.default.TSS.upstream}}{Define the upstream range of transcription start site region.}
 \item{\code{.default.TSS.downstream}}{Define the downstream range of transcription start site region.}
 \item{\code{.default.max.flankWindow}}{Define the max size of the flanking window surrounding the 
 predicted RE-CpG.}
 \item{\code{.default.450k.annotation}}{A character string associated with the Illumina 450k 
 array annotation dataset.}
 \item{\code{.default.epic.annotation}}{A character string associated with the Illumina EPIC 
 array annotation dataset.}
 \item{\code{.default.genomicRegionColNames}}{Define the names of the genomic regions for prediction.}
 \item{\code{.default.predictors}}{Define the names of predictors for RE methylation prediction.}
 \item{\code{.default.mtry.tune}}{Define the default \code{mtry} parameter for Random Forest model.}
 \item{\code{.default.C.svmLinear.tune}}{Define the default \code{C} (Cost) parameter for Support 
 Vector Machine (SVM) using linear kernel.}
 \item{\code{.default.sigma.svmRadial.tune}}{Define the default \code{sigma} parameter for SVM
 using Radial basis function kernel.}
 \item{\code{.default.C.svmRadial.tune}}{Define the default \code{C} (Cost) parameter for SVM
 using Radial basis function kernel.} 
}
}

\examples{
# Display all default settings
remp_options()

# Change default maximum flanking window size to 2000
remp_options(.default.max.flankWindow = 2000)

# Reset all options
remp_reset()

}
