% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remp.R
\name{remp}
\alias{remp}
\title{Repetitive element methylation prediction}
\usage{
remp(methyDat, REtype = c("Alu", "L1"), parcel = NULL,
  work.dir = tempdir(), groom = TRUE, win = 1000, method = c("rf",
  "svmLinear", "svmRadial", "naive"), autoTune = TRUE, param = NULL,
  seed = NULL, ncore = NULL, BPPARAM = NULL, verbose = FALSE)
}
\arguments{
\item{methyDat}{A \code{\link{RatioSet}}, \code{\link{GenomicRatioSet}}, \code{\link{DataFrame}}, 
\code{data.table}, \code{data.frame}, or \code{matrix} of methylation dataset. See Details.}

\item{REtype}{Type of RE. Currently \code{"Alu"} and \code{"L1"} are supported.}

\item{parcel}{An \code{\link{REMParcel}} object containing necessary data to carry out the
prediction. If \code{NULL}, the function will search the \code{.rds} data file in \code{work.dir} 
exported by \code{\link{initREMP}} (with \code{export = TRUE}) or \code{\link{saveParcel}}.}

\item{work.dir}{Path to the directory where the annotation data generated by \code{\link{initREMP}} 
are saved. Valid when the argument \code{parcel} is missing. If not specified, temporary directory 
\code{tempdir()} will be used. If specified, the directory path has to be the same as the 
one specified in \code{\link{initREMP}} or in \code{\link{saveParcel}}.}

\item{groom}{Should the function run \code{\link{grooMethy}} implicitly to check and fix the data? 
Default = \code{TRUE}. If \code{\link{grooMethy}} has been run in advance, let \code{groom = FALSE} can 
avoid repeated check.}

\item{win}{An integer specifying window size to confine the upstream and downstream flanking 
region centered on the predicted CpG in RE for prediction. Default = \code{1000}. See Details.}

\item{method}{Name of model/approach for prediction. Currently \code{"rf"} (Random Forest),
\code{"svmLinear"} (SVM with linear kernel), \code{"svmRadial"} (SVM with linear kernel), and
\code{"naive"} (carrying over methylation values of the closest CpG site) are available. 
Default = \code{"rf"} (Random Forest). See Details.}

\item{autoTune}{Logical parameter. If \code{TRUE}, a 3-time repeated 5-fold cross validation 
will be performed to determine the best model parameter. If \code{FALSE}, the \code{param} option 
(see below) must be specified. Default = \code{TRUE}. Auto-tune will be disabled using Random Forest. 
See Details.}

\item{param}{A number or a vector specifying the model tuning parameter(s) (not applicable for Random Forest).
For SVM, \code{param} represents 'Cost' (for linear kernel) or 'Sigma'
and 'Cost' (for radial basis function kernel). This parameter is valid only when \code{autoTune = FALSE}.}

\item{seed}{Random seed for Random Forest model for reproducible prediction results. 
Default is \code{NULL}, which generates a seed.}

\item{ncore}{Number of cores to run parallel computation. By default, max number of cores available 
in the machine will be utilized. If \code{ncore = 1}, no parallel computation is allowed.}

\item{BPPARAM}{An optional \code{\link{BiocParallelParam}} instance determining the parallel back-end to 
be used during evaluation. If not specified, default back-end in the machine will be used.}

\item{verbose}{Logical parameter. Should the function be verbose?}
}
\value{
An \code{\link{REMProduct}} object containing prediction results.
}
\description{
\code{remp} is used to predict genomewide methylation levels of locus-specific repetitive elements (RE).
Two major RE types in human, Alu element (Alu) and LINE-1 (L1) are available.
}
\details{
Before running \code{remp}, user should make sure the methylation data have gone through 
proper quality control, background correction, and normalization procedures. Both beta value
and M value are allowed. Rows represents probes and columns represents samples. Please make 
sure to have row names that specify the Illumina probe ID (i.e. cg00000029). Parameter 
\code{win = 1000} is based on previous findings showing that neighboring CpGs are more likely 
to be co-modified within 1000 bp. User can specify narrower window size for slight improvement of 
prediction accuracy at the cost of less predicted RE. Window size greater than 1000 is not
recommended as the machine learning models would not be able to learn much userful information 
for prediction but introduce noise. Random Forest model (\code{method = "rf"}) is recommented 
as it offers more accurate prediction and it also enables prediction reliability functionality. 
Prediction reliability is estimated by conditional standard deviation using Quantile Regression Forest. 
Please note that if parallel computing is allowed, parallel Random Forest 
(powered by package \code{\link{ranger}}) will be used automatically. The performance of 
Random Forest model is often relatively insensitive to the choice of \code{mtry}.
Therefore, auto-tune will be turned off using Random Forest and \code{mtry} will be set to one third 
of the total number of predictors. For SVM, if \code{autoTune = TRUE}, preset tuning parameter 
search grid can be access and modified using \code{\link{remp_options}}.
}
\examples{
# Obtain example Illumina example data (450k)
GM12878_450k <- getGM12878('450k')

# Make sure you have run 'initREMP'. See ?initREMP.

# Run prediction (grooMethy will be run implicitly)
remp.res <- remp(GM12878_450k, REtype = 'Alu', ncore = 1)
remp.res
details(remp.res)

# Extract CpG location information (inherit from class 'RangedSummarizedExperiment')
rowRanges(remp.res) 

# RE annotation information
annotation(remp.res)

# Add gene annotation
remp.res <- decodeAnnot(remp.res, type = "symbol", ncore = 1)
annotation(remp.res)

# Density plot across predicted CpGs in RE
plot(remp.res)

# Trim off less reliable prediction
remp.res <- trim(remp.res)
plot(remp.res)

}
\seealso{
See \code{\link{initREMP}} to prepare necessary annotation database before running \code{remp}.
}
