% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMPtools.R
\name{remprofile}
\alias{remprofile}
\title{Extract DNA methylation data profiled in RE}
\usage{
remprofile(methyDat, REtype = c("Alu", "L1"), RE = NULL, verbose = FALSE)
}
\arguments{
\item{methyDat}{A \code{\link{RatioSet}}, \code{\link{GenomicRatioSet}}, \code{\link{DataFrame}},
data.table, data.frame, or matrix of Illumina BeadChip methylation data (450k or EPIC array).}

\item{REtype}{Type of RE. Currently \code{"Alu"} and \code{"L1"} are supported.}

\item{RE}{A \code{\link{GRanges}} object containing user-specified RE genomic location information. 
If \code{NULL}, the function will retrive RepeatMasker RE database from \code{\link{AnnotationHub}} 
(build hg19).}

\item{verbose}{Logical parameter. Should the function be verbose?}
}
\value{
A \code{\link{REMProduct}} object containing profiled RE methylation results.
}
\description{
\code{remprofile} is used to extract profiled methylation of CpG sites in RE.
}
\examples{
data(Alu.demo)
GM12878_450k <- getGM12878('450k')
remp.res <- remprofile(GM12878_450k, REtype = "Alu", RE = Alu.demo)
details(remp.res)
rempB(remp.res) # Methylation data (beta value)

remp.res <- rempAggregate(remp.res)
details(remp.res)
rempB(remp.res) # Methylation data (beta value)

}
