% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipp_wm_wide.R
\docType{data}
\name{tipp_wm_wide}
\alias{tipp_wm_wide}
\title{Tipp WM (breites Format)}
\format{
Ein tibble mit 64 Zeilen and 7 Spalten:
\describe{
  \item{Thomas_Bayes}{Tipps von Thomas Bayes.}
  \item{Egon_Pearson}{Tipps von Egon Pearson.}
  \item{Jerzy_Neyman}{Tipps von Jerzy_Neyman.}
  \item{Alice}{Tipps von Alice.}
  \item{Bob}{Tipps von Bob.}
  \item{Charlie}{Tipps von Charlie.}
  \item{Ergebnis}{Wahres Ergebnis des Spiels.}
}
}
\usage{
tipp_wm_wide
}
\description{
Tipps von sechs verschiedenen Personen zur Fußball Weltmeisterschaft 2018 in breitem Format.
}
\examples{
tipp_wm_wide
}
\keyword{datasets}
